+++ src/Network/Network/ChangeLog.md
@@ -19,6 +19,10 @@
 --->
 
 ## Upcoming Release
+* Added support to create, get and delete Bastion shareable links
+    - `New-AzBastionShareableLink`
+    - `Get-AzBastionShareableLink`
+    - `Remove-AzBastionShareableLink`
 * Added support of `UserAssignedIdentityId` Property in `New-AzNetworkWatcherFlowLog` and `Set-AzNetworkWatcherFlowLog` commands
 
 ## Version 7.8.0